+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Official support for unpatched Mapbox GL version 0.50
 - Add `isNull` supporting numerical and categorical properties, deprecating `isNaN()`
-- Change type from `viewportFeatures` expression, adding vizProperties to features as in `Interactivity`
-- Add `filter` and `transform` properties to features return from `Interactivity` events.
+- Add `filter` and `transform` properties to features returned from `Interactivity` events.
+- Change type from `viewportFeatures` expression, adding vizProperties and variables to features as in `Interactivity`
 
 ### Changed
 - Eliminate the parameter to specify a property in `globalCount()` and `viewportCount()`