+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix `viewportPercentile` (styling by it was broken)
 - Fix `viewport*` functions to take clustering into account
 - Fix wrong `expressionName` due to mangled `this.constructor.name` in the minified version
+- Fix a regression error related to `blendTo` when using feature `reset`.
 
 ## [0.9.1] - 2018-10-09
 ### Fixed